+++ tests/test_search.py
@@ -66,27 +66,6 @@ def test_get_dashboard__other_folder(ro_demo_grafana):
 	assert r["title"] == "f0-0"
 
 
-def test_resolve_path__general():
-	assert Finder._resolve_path("/folder/dashboard") == ("folder", "dashboard")
-
-
-def test_resolve_path__no_absolute_slash():
-	assert Finder._resolve_path("folder/dashboard") == ("folder", "dashboard")
-
-
-def test_resolve_path__implicit_general():
-	assert Finder._resolve_path("/dashboard") == ("General", "dashboard")
-
-
-def test_resolve_path__bare_dashboard():
-	assert Finder._resolve_path("dashboard") == ("General", "dashboard")
-
-
-def test_resolve_path__too_many_parts():
-	with pytest.raises(ValueError):
-		Finder._resolve_path("/folder/dashboard/invalidpart")
-
-
 def test_get_from_path__smoke(ro_demo_grafana):
 	f = Finder(ro_demo_grafana[1])
 	r = f.get_from_path("/f0/f0-0")